beat the 500 line allegations
